+++ src/components/forms/fields/index.ts
@@ -3,6 +3,8 @@ export * from './FormError';
 export * from './FormInput';
 export * from './FormSelect';
 export * from './FormCheckbox';
+export * from './FormCheckboxGroup';
+export * from './FormBadgePicker';
 export * from './FormDropdown';
 export * from './FormDateRange';
 export * from './FormBadgesSelection';